LayoutTests/fast/forms/datetimelocal/datetimelocal-editable-components/datetimelocal-editable-components-mouse-events.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/common.js"></script>
<script src="../../../../resources/js-test.js"></script>
<script src="../../../../resources/ui-helper.js"></script>
<style>
input {
    width: 600px;
    font-size: 30px;
}
</style>
</head>
<body>

<input id="input" type="datetime-local" value="2020-12-20T05:30">

<script>

description("Test for mouse events for &lt;input type=datetime-local&gt;");

clickEventsFired = 0;
function onClickEvent() {
    clickEventsFired++;
}

function mouseClickOn(x, y) {
    if (!window.eventSender)
        return;
    eventSender.mouseMoveTo(x + input.offsetLeft, y + input.offsetTop);
    eventSender.mouseDown();
    eventSender.mouseUp();
}

input.addEventListener("click", onClickEvent);
const center = input.offsetHeight / 2;

const shadowRoot = internals.shadowRoot(input);
const monthField = getElementByPseudoId(shadowRoot, "-webkit-datetime-edit-month-field");
const dayField = getElementByPseudoId(shadowRoot, "-webkit-datetime-edit-day-field");
const yearField = getElementByPseudoId(shadowRoot, "-webkit-datetime-edit-year-field");
const hourField = getElementByPseudoId(shadowRoot, "-webkit-datetime-edit-hour-field");
const minuteField = getElementByPseudoId(shadowRoot, "-webkit-datetime-edit-minute-field");
const meridiemField = getElementByPseudoId(shadowRoot, "-webkit-datetime-edit-meridiem-field");

debug("Enabled Input\n");

UIHelper.activateElement(monthField);
UIHelper.keyDown("9");
shouldBeEqualToString("input.value", "2020-09-20T05:30");

UIHelper.activateElement(dayField);
UIHelper.keyDown("1");
UIHelper.keyDown("2");
shouldBeEqualToString("input.value", "2020-09-12T05:30");

UIHelper.activateElement(yearField);
UIHelper.keyDown("3");
UIHelper.keyDown("0");
UIHelper.keyDown("3");
UIHelper.keyDown("0");
shouldBeEqualToString("input.value", "3030-09-12T05:30");

UIHelper.activateElement(hourField);
UIHelper.keyDown("8");
shouldBeEqualToString("input.value", "3030-09-12T08:30");

UIHelper.activateElement(minuteField);
UIHelper.keyDown("4");
UIHelper.keyDown("5");
shouldBeEqualToString("input.value", "3030-09-12T08:45");

UIHelper.activateElement(meridiemField);
UIHelper.keyDown("P");
shouldBeEqualToString("input.value", "3030-09-12T20:45");

// Click on control, but not a specific field, defaults to first field.
mouseClickOn(550, center);
UIHelper.keyDown("6");
shouldBeEqualToString("input.value", "3030-06-12T20:45");

// Click outside control.
mouseClickOn(input.offsetWidth + 5, input.offsetHeight + 5);
UIHelper.keyDown("5");
shouldBeEqualToString("input.value", "3030-06-12T20:45");

shouldBe("clickEventsFired", "7");

debug("\nDisabled Input\n");
clickEventsFired = 0;
input.disabled = true;
input.readOnly = false;

UIHelper.activateElement(monthField);
UIHelper.activateElement(dayField);
UIHelper.activateElement(yearField);
UIHelper.activateElement(hourField);
UIHelper.activateElement(minuteField);
UIHelper.activateElement(meridiemField);
// Click on control, but not a specific field, defaults to first field.
mouseClickOn(550, center);
// Click outside control.
mouseClickOn(input.offsetWidth + 5, input.offsetHeight + 5);

shouldBe("clickEventsFired", "0");

debug("\nReadonly Input\n");
clickEventsFired = 0;
input.disabled = false;
input.readOnly = true;

UIHelper.activateElement(monthField);
UIHelper.activateElement(dayField);
UIHelper.activateElement(yearField);
UIHelper.activateElement(hourField);
UIHelper.activateElement(minuteField);
UIHelper.activateElement(meridiemField);
// Click on control, but not a specific field, defaults to first field.
mouseClickOn(550, center);
// Click outside control.
mouseClickOn(input.offsetWidth + 5, input.offsetHeight + 5);

shouldBe("clickEventsFired", "7");

</script>
</body>
</html>
